requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

sqlalchemy==2.0.34
psycopg2-binary==2.9.9

pillow==10.4.0
numpy==2.1.1
faiss-cpu==1.8.0.post1
opencv-python-headless==4.10.0.84

insightface==0.7.3
onnxruntime==1.19.2
deepface==0.0.93

python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0

cloudinary


fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
requests
